message_summary: Factor out into the actions package.
